Avancement fonction + debuggage

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: vgroux <vgroux@student.42lausanne.ch>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/10 16:37:27 by vgroux            #+#    #+#             */
-/*   Updated: 2022/10/10 19:09:48 by vgroux           ###   ########.fr       */
+/*   Updated: 2022/10/11 13:13:56 by vgroux           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
